ux updates to lu editor placeholders
- make editors 50% larger
- add special placeholder for expected responses
