+++ Composer/packages/client/src/components/ProjectTree/TriggerCreationModal.tsx
@@ -15,8 +15,8 @@ import { TextField } from 'office-ui-fabric-react/lib/TextField';
 import { luIndexer, combineMessage } from '@bfc/indexers';
 import { PlaceHolderSectionName } from '@bfc/indexers/lib/utils/luUtil';
 import get from 'lodash/get';
-import { DialogInfo, inlineModePlaceholder } from '@bfc/shared';
-import { LuEditor } from '@bfc/code-editor';
+import { DialogInfo } from '@bfc/shared';
+import { LuEditor, inlineModePlaceholder } from '@bfc/code-editor';
 
 import {
   generateNewDialog,
@@ -262,20 +262,22 @@ export const TriggerCreationModal: React.FC<TriggerCreationModalProps> = props =
               data-testid={'RegExDropDown'}
             />
           )}
-          {showTriggerPhrase && <Label>{formatMessage('Trigger phrases')}</Label>}
           {showTriggerPhrase && (
-            <LuEditor
-              onChange={onTriggerPhrasesChange}
-              value={formData.triggerPhrases}
-              errorMessage={formData.errors.triggerPhrases}
-              luOption={{
-                projectId,
-                fileId: dialogId,
-                sectionId: formData.intent || PlaceHolderSectionName,
-              }}
-              height={150}
-              placeholder={inlineModePlaceholder}
-            />
+            <React.Fragment>
+              <Label>{formatMessage('Trigger phrases')}</Label>
+              <LuEditor
+                onChange={onTriggerPhrasesChange}
+                value={formData.triggerPhrases}
+                errorMessage={formData.errors.triggerPhrases}
+                luOption={{
+                  projectId,
+                  fileId: dialogId,
+                  sectionId: formData.intent || PlaceHolderSectionName,
+                }}
+                height={225}
+                placeholder={inlineModePlaceholder}
+              />
+            </React.Fragment>
           )}
         </Stack>
       </div>
